requirements.txt

streamlit
faiss-cpu
pyyaml
openai
truststore
